+++ astropy/units/tests/test_quantity.py
@@ -6,6 +6,8 @@ import copy
 import pickle
 import decimal
 import numbers
+import sys
+import typing as T
 from fractions import Fraction
 
 import pytest